agentic-orchestration-tool/requirements.txt
crewai
python-dotenv
pyyaml
httpx
